+++ src/spec-common/injectHeadless.ts
@@ -62,9 +62,9 @@ export interface ResolverParameters {
 	buildxOutput: string | undefined;
 	skipFeatureAutoMapping: boolean;
 	skipPostAttach: boolean;
-	experimentalImageMetadata: boolean;
 	containerSessionDataFolder?: string;
 	skipPersistingCustomizationsFromFeatures: boolean;
+	omitConfigRemotEnvFromMetadata?: boolean;
 }
 
 export interface LifecycleHook {