+++ CHANGELOG.md
@@ -2,7 +2,34 @@
 
 Notable changes.
 
+## April 2023
+
+### [0.38.0]
+
+- Update vm2 (https://github.com/devcontainers/cli/pull/488)
+
+### [0.37.0]
+
+- Add --config to build command (microsoft/vscode-remote-release#8068)
+- Features/Templates: Fix a bug in reading from docker credential helpers (https://github.com/devcontainers/cli/issues/477)
+
 ## March 2023
+
+### [0.36.0]
+
+-  Add initial support for docker credential helpers when fetching Features/Templates. (https://github.com/devcontainers/cli/pull/460, contributed by @aaronlehmann)
+
+### [0.35.0]
+
+- Transform maven, gradle and jupyterlab usages to their features v2 counterparts. (https://github.com/devcontainers/cli/issues/461)
+- Escape and enclose containerEnv in quotes when writing to Dockerfile. (https://github.com/devcontainers/cli/issues/454)
+- Update package dependencies.
+
+### [0.34.0]
+
+- Also require name property in `devcontainer-feature.json`. (https://github.com/devcontainers/cli/pull/447)
+- Add `--omit-config-remote-env-from-metadata` to omit remoteEnv from devcontainer config on container metadata label. (https://github.com/devcontainers/cli/pull/453)
+- Only include required legacy scripts. (https://github.com/microsoft/vscode-remote-release/issues/7532)
 
 ### [0.33.0]